requirements.txt
# Core Discord Bot Dependencies
discord.py>=2.3.0
python-dotenv>=1.0.0

# AI and Analysis Dependencies
openai>=1.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
requests>=2.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Dashboard Dependencies
Flask>=2.3.0
Flask-CORS>=4.0.0

# Utilities
python-dateutil>=2.8.0
psutil>=5.9.0

# Additional Dependencies for Enhanced Features
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# New dependencies
aiofiles>=23.2.1
